requirements.txt
pytest-playwright == 0.4.4
pytest-xdist == 3.5.0
pytest-html==4.1.1